feat: add overarching cli
