+++ visgpu.yaml
@@ -1,4 +1,4 @@
-simulator: VisCPU
+simulator: MatVis
 precision: 2
 ref_time: mean
 correct_source_positions: true
